full rewrite for async/await (#66)
This is a full rewrite the crate.

It provides both sync. decoding (parsing) and encoding (writing) and async operations.

The default is async using the Client. The old iterator/filter thing is gone, now its simply just 'event subscriptions'  on a dispatcher. This returns an async Stream of specific Messages.

The big enum of message types is gone. 

Message types can be owned or borrowed from an input string. There are traits to allow conversions between them.

For example: 
* the Join message is defaulted to a Join<String> (the owned representation) but as_borrowed() can be called on it to get it as a Join<&str>. The opposite is also true
* Message<&str> can be converted to a Message<String> with an into()/from()
* indv. messages (Join, et.al) can be 'parsed' from a Message<T> by using the Parse trait.
* The default 'storage' type for these messages (the generic param) is 'String'

Event subscription:

The client has a 'Dispatcher' which lets you subscribe to specific events.
for example: client.dispatcher().await.subscribe::<event::Join>()
This will return an EventStream of Arc<messages::Join> which you can get via the futures Stream behavior.

The dispatcher allows you clear all subscriptions for a specific event, or all events.

When it cannot send to the EventStream, it'll mark it as dead and the next message for that event will remove any dead subscribers.
You should consume an EventStream to completion -- or drop it to ensure the dispatcher cleans up these old subscriptions.

The client, and its writer are both Clonable (and Send+Sync) so they can be sent across threads/tasks.

When the client has stop() called on it, it'll clear subscriptions and start a graceful shutdown.

The client takes an AsyncRead and AsyncWrite in its run() method which'll return a future to the result of the client's operation. e.g. Canceled, EOF or an Error.

This is the first version to actually support TLS.