Add an extract trait for AllCommands.
This is an alternative to using a match statement. This takes a borrowed
AllCommands and produces a Option<&T>.

This implements a suggestion from #151. Originally I was going to use
the TryFrom/TryInto traits but they are tedious. (i.e. Result<&T, ()>)

The trait isn't exported into the root scope because its rather ..
specific but I've needed it for another use of the crate. So, to access
it, you need to use twitchchat::messages::Extract as _; There is no
owned version of this, the conversion just reborrow as the inner type.
Technically the Reborrow trait /could/ do this, but this can fail if the
AllCommands was not the provided shape.